OSX: fix compiler warning

+++ psutil/_psutil_osx.c
@@ -338,7 +338,10 @@ psutil_proc_memory_maps(PyObject *self, PyObject *args) {
             memset(addr_str, 0, sizeof(addr_str));
             memset(perms, 0, sizeof(perms));
 
-            sprintf(addr_str, "%016lx-%016lx", address, address + size);
+            sprintf(addr_str,
+                    "%016lx-%016lx",
+                    (long unsigned int)address,
+                    (long unsigned int)address + size);
             sprintf(perms, "%c%c%c/%c%c%c",
                     (info.protection & VM_PROT_READ) ? 'r' : '-',
                     (info.protection & VM_PROT_WRITE) ? 'w' : '-',
